Added method for muon constraint to vertex before matching
By default the constraint is on (with val=0.1), det->SetApplyBransonPCorrection(val) will enforce constraint if
val>=0 and disable it otherwise. The value of val is added to vertex errors as additional tolerance, to not lose
offsetted muons.
Also committed the files/KMCUtils.cxx with solenoid field corresponding to long setup