Interface to Fluka (Enrico)
Original Enrico's message:
after some days of work I converged in preparing an interface that allows one to
use the GenMUONLMR class and all its signal generation options  as an input to fluka.
I send you a tar archive containing all the needed files. I tested and debugged it
with the various generation options and, as far as I can see, everything works