Magner can be set from the setup (Alessandro de Falco)
Message from Alessandro: I changed a bit the classess KMCDetectorForw and KMCUtils such that the magnets configuration can be read from the setup file, rather than changing the hardcoded values.
The modification of the classes is backward compatible, so the old setup files still work.

I also wrote a macro that displays the setup (in two ways, one that just plots boxes, and another one that creates a geometry based on TGeo.
I attach the relevant code and also a setup file containing the description of the magnets.