Modification from Enrico
Original message:
1) GenMUONLMRLinkDef.h (in Fluka directory)

This was plainly missing in the previous push and is needed for the external compilation of the GenMUONLMR class used by the Fluka interface

2) GenMUONLMR.cxx, GenMUONLMR.h (in the files directory)

Here I modified the classes recently updated by Alessandro, in order to have the jpsi being generated. Two possible parameterizations are available, Pythia and Schuler. Note that I had a private copy of the class in the Fluka directory, but this is impractical for the future if more developments on the class itself will be needed. Consequently, the current GenMUONLMR.cpp, GenMUONLMR.h in the Fluka directory should be removed.

The only non-backward compatible modification is that I had to add one argument to the SetYParams() method. I checked that this is only used as of today by the macro runDiMuGenLMR.C. A zero should be added as last argument there.

3) Flulib.cpp (in the Fluka directory)

This is the interface C/F77 file and was also modified to allow generation of J/psi with various parameterizations.

4) Makefile (in Fluka directory)

This is now modified, mainly to use the GenMUONLMR.cxx, GenMUONLMR.h in the parent files directory

5) NA60plus-lowenergy.inp (in Fluka directory)

This is the Fluka set-up used in the simulations for the EoI, I believe it is better to have it in a safe place

6) rfluka.sh (in Fluka directory)

Slightly modified to work with the committed set-up file