New setups from Enrico, original message below
we had a discussion last week on the  set-up with the AMS magnet and we agreed to prepare some other set-up files, corresponding to a situation where only part of the AMS field is used (60 cm), and 7 pixel planes are considered.

I prepared those, plus some others corresponding to various field values in a 40 cm space with 5 pixel planes. These set-ups use the improvement introduced by Alessandro for defining the fields inside the set-up files.