Finalize fluka background import to KMCDetector
runDiMuGenLMR.C is able to overlay fluka background on top of the generated signal,
provided that flukaBGList argument is non-empty and points on the text file with fluka
dumps.
interactionSource argumnent is an optional string to select primary interaction volume
(e.g. "target" will select only events with interaction in volumes "target..."
If there are no enough background events to guarantee generation of requested number of signal
events, the fluka parser will wrap to the beginning of the list.
As for the parametric background, the refreshBg argument tells per how many signal events the
background should be updated.

If flukaBGList is an empty string, the parametric background will be used.