Optionally impose no bg.clusters pickup at last ITS layer
If det.setLastITSLayerClean(true); is called on the instanciated KMCDetectorFwd detector instance,
then no background clusters as well as signal inefficiency will be considered in the tracking
on the last ITS layer.
Therefore, if the track-to-cluster tolerances are large enough and the signal cluster was in the acceptance
of the layer, it will be attached to the track. Otherwise, this layer will not contribute to the track at all
(and the reconstruction of the track will fail instead of producing a fake,
if the contribution from the last layer was requested for the track validation).